pyproject.toml
[project]
name = "icd-chapter-mistral"
version = "0.1.0"
description = "Fine-Tuning Mistral models for predicting ICD Chapters of Discharge Notes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.0.1",
    "matplotlib==3.10.7",
    "medspacy>=1.3.1",
    "numpy>=2.0.0",
    "pandas>=2.3.3",
    "scipy>=1.16.2",
    "wandb>=0.18.0",
    "scikit-learn>=1.5.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.66.0",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "mistralai>=1.9.11",
    "transformers>=4.57.1",
    "sentencepiece>=0.2.1",
]
